+++ transformer_engine/pytorch/attention/__init__.py
@@ -5,7 +5,7 @@
 """Python interface for attention"""
 
 from .dot_product_attention import DotProductAttention
-from .linear_attention import LinearAttention
+from .linear_attention import GatedDeltaNetAttention
 from .fused_mla_q_uproj import FusedMLAQUpProjFunction, FusedMLAQUpProjRopeQuant
 from .multi_head_attention import MultiheadAttention
 from .inference import InferenceParams
@@ -13,7 +13,7 @@ from .rope import RotaryPositionEmbedding
 
 __all__ = [
     "DotProductAttention",
-    "LinearAttention",
+    "GatedDeltaNetAttention",
     "FusedMLAQUpProjFunction",
     "FusedMLAQUpProjRopeQuant",
     "MultiheadAttention",
